Moved and renamed SignInValidationTest, SignUpValidationTest; Refactored these classes
